Fix device set for values having spaces
This patch fixes the device set command when there are values
which contain spaces.
For example,

The command pvr device set <my_nice> my_key="a spaced value"
would only be set as "a" while the rest of the part would be
discarded.

This is required for example when trying to add a key viz
pvr-sdk.authorized_keys which can contain spaces.

Signed-off-by: Pranay Kumar Srivastava <pranay.srivastava@pantacor.com>